Add VKontakte OneTap auth and adjust styles
This commit introduces the VKontakte OneTap authentication to the application. A new block with class ".auth" is added to the main layout to accommodate the authentication iframe. Additional CSS styles are applied to this block and the main element for a better visual integration of the auth feature. The necessary scripts for OneTap auth and its configuration are added in "main.mjs". ".gitignore" file is also updated to ignore node_modules, package files, and .vercel directory. Two new files ("api/index.mjs" and ".gitignore") are also added. Updating the .gitignore file prevents unnecessary project files from being tracked by Git.