"Update title, scripts, styles and functionality in index.html, main.css and main.mjs"
This commit makes several changes to different files. In index.html, the title was simplified by removing the '300' at the end. Some scripts were put at the end of the body tag instead of being in the head. The main.css file had a small adjustment to place an item closer to the bottom.

Most of the changes are in main.mjs where the script was refactored to streamline the logic related to the OneTap button and alter the appearance and behavior based on newly understood requirements and user feedback. The imports were also moved to be more aligned with commonly used practices and increase code readability.