Update project dependencies and create new HTML components
This commit updates the package-lock.json and package.json to properly move the '@vkontakte/superappkit' package from devDependencies to Dependencies as it is needed in the production build. A new version of 'lit-edge-utils' from PonomareVlad is used. Newly created HTML elements (app.mjs, client.mjs, and page.mjs) are added to the assets directory. The corresponding scripts are inserted in the api/index.mjs file and importmap.json is created for mapping the imported packages. This is needed for better component organization and enhanced readability of the code.