Refactor voting functionality and update dependencies
Refactored vote-related HTTP request processing by moving most of the logic to the api. This was done to improve code cleanliness, separations of concerns and ensure better testability. Now vote function checks the validity of parameters, finds the target round and updates the vote accordingly. Also, added a new dependency 'bson' and updated the existing ones in package.json and package-lock.json for keeping the project dependencies up-to-date and ensure compatibility with latest standards.