Refactor auth and vote functionality for improved structure and efficiency
This commit involved several changes to improve the structure and efficiency of the auth and vote functionality. The auth logic has been moved to a separate auth function in the api.mjs file. This makes the code cleaner and easier to maintain. In the app.mjs file, the new auth function is used and error handling has been enhanced. The implementation of vote functionality has also been improved by aborting previous fetch requests when a new vote is triggered. This should result in more responsive voting. Also updated the retry logic for updating account and round state, as well as included better abort signal handling for various asynchronous tasks. Additionally, minor changes to the app.styles.mjs file were made to improve the visual feedback to the user.