Update User Interface and Improve Number Formatting
Updated the format in which numbers are displayed to the users by introducing a new function for number formatting, Intl.NumberFormat. This improvement results in fraction-based numbers being displayed with no more than 1 decimal, a formatting style suitable for Russian users, as specified in the requirements.

On the User Interface side, significant enhancements were made to the voting process. A new design was introduced for the vote section, providing a more interactive experience for users who can now participate in voting, see a success message upon voting, and be able to follow the voting results online. This improvement should introduce a positive impact on the user experience of the application.

Changes also include renaming the class "auth" to "vk" and updating inline styles for it. Other style modifications were applied to elements such as buttons, page layout, images, background...etc.

In conclusion, this commit comprises a significant update in the application's look and feel, and operation, hence enhancing users' interaction with it.