Remove old CSS files
Deleted main.css and vars.css from the old directory. The styles in these files were outdated and no longer needed, as we've transitioned to a new design. Added touch-action and overflow styles to root element in app.css, and transform to the logo, to enhance user's interface interaction and improve performance.