Refactor media queries in vars.css
The landscape orientation and portrait orientation styles were split into separate media queries. This was done to improve readability and support different devices. The min-width of the orientation was increased from 700px to 750px for better optimization. Also, a new media query was added for landscape orientation with max-height of 500px for flexibility on smaller devices.