Separate background image into left and right for decorations
In order to handle different background images for the left and right of the decoration section, we separated the `--decorations-background-image` into `--decorations-background-left-image` and `--decorations-background-right-image`. Also added a new variable `--vote-decorations-background-image` to handle alternative right decoration for the voting section. All these changes are applied in vars.css and app.css providing increased control over the visualization of the website.