Update CSS for better responsiveness
Changed the width and horizontal padding of the vote section and variants in vars.css, and applied these changes in app.css. These adjustments were necessary to improve the layout and responsiveness of the website on different screen sizes. Removed hard-coded width in app.css to utilize CSS variables for easy modifications in the future.