Remove redundant code and update session management logic
Removed "mask-icon" link tag from the document head in page.mjs as it was unnecessary.

In client.mjs, rename the const logtail to logger, so that it can be referred to as a logger globally on the window object.

In app.mjs, Added a console Proxy as logger to replicate all console-level logs to Logtail logs. Updated the vote function with a new object structure to support multiple voting per user per round, stored in the account object. Also added a scheduling function for updateRoundState to give users real-time updates without refreshing the page.

In api.mjs, updated the auth and vote functions to support multiple voting per user per round and to give more descriptive error messages and cleaner variable usage. Also updated the response structure of the auth function to return a complete list of user choices for each round.

In api/index.mjs & api/round.mjs changed cache option to 1 second to enable almost real-time update after each vote.

Finally, removed the date property from the client-side state as it is already included in the round data in the server-side getLastRound function.