Refactor vote function and query for active rounds
Changes are made in the vote function in app.mjs to pass round_id directly instead of extracting from this.round object. This ensures the current round_id being used is the intended one.

Exported activeRoundQuery is also created in api.mjs to minimize redundancy and increase maintainability. This query is to fetch only incomplete rounds. This change increases the readability and maintainability of the code.