Update domain name in config.json
The domain name in the config.json file was changed from "rcc-vote.vercel.app" to "rcc-vote.ru". This update is made due to the migration of the application from the Vercel platform to a separate domain under ".ru". This change will ensure the correct connection of the application with the new domain.