Update image srcset links in app.mjs
The source set (srcset) links in app.mjs were pointed from the previous domain "rcc-vote-images.vercel.app" to the new domain "cdn.rcc-vote.ru". This was necessitated due to the shift of the application host from the Vercel platform to a custom ".ru" domain. The correct links are crucial to maintaining smooth loading and viewing of images on different screen resolutions.