Update CSS styles for better UI experience
Multiple changes were made to the CSS files to improve the overall user interface. The 'background' property was changed to 'background-color' for better browser compatibility. Furthermore, properties representing padding values were revised and minor changes were introduced to ensure alignment and positioning of elements. To improve visibility 'font-size' was added to the button class. The layout structure of the 'logo' class was reset to center using flex and the 'transform-origin' of '.logo img' was set to center top. The '.subscribe' class was refactored to have '.controls' for a better control group. Resolution queries and color values were tweaked for a smoother responsive design. Lastly, HTML structure in 'app.mjs' was modified to organize the controls better and gear towards a more intuitive design. All these updates were aimed at making the website more user-friendly and visually appealing.