Refactor stylesheets and alter layout to improve user interface
The website's styles and layout were revised to enhance its appearance and accessibility. Various classes in the CSS were updated, with superfluous styles being removed. A significant modification was made to auth.css, which was moved to the views folder for better organization. Another necessary change was the addition of "overflow:hidden" to the CSS, which solves the scrolling issue on the webpage. App.mjs was updated as well to perform authentication-specific features. Changes overall should improve the user interface and experience on the site.