Add dynamic countdown timer in the app
The commit added a dynamic countdown timer feature in the app. 'ifDefined' directive was imported in app.mjs, with the timestamp property being passed as an attribute to app-countdown element. The 'countdown' attribute was also added to the 'app-root' element in api/index.mjs with its value fetched from the environment variables.

In public/countdown.mjs, a countdown functionality was added by creating a timer that calculates and updates the days, hours, minutes, and seconds until a certain timestamp recalculating every 1000 milliseconds. The timer is cleared to avoid unnecessary computation when the component is disconnected.

The addition of this feature allows events to display the remaining time until they start, therefore enhancing the user experience.