Optimize image loading by using lazy loading
Ensuring a smooth user experience on our site involves reducing load times where possible. In the diff, the 'loading' attribute for the image element was set to 'lazy'. This change was made to optimize image loading, reducing the amount of data that needs to be loaded when the page first loads, and delaying loading of images until they are needed.